requirements.txt
boto3==1.17.3
PyGithub==1.54.1
requests==2.25.1
